requirements.txt
# FastAPI + ASGI server
fastapi
uvicorn[standard]

# OpenAI API client
openai

# PostgreSQL DB driver
psycopg[binary]   # Better than psycopg2, works well with PostgreSQL 15

# JWT Authentication
PyJWT

# Pydantic for data models
pydantic

# Optional: environment variable management
python-dotenv

# Optional: if you decide to use pgvector for semantic search
# pgvector
PyYAML

# testing 
pytest
httpx
